storage/filesystem: make the driver work on Windows
- Close the writer before renaming in PutContent: Windows refuses to
  rename a file while a handle to it is still open.
- Treat storagedriver keys as slash-separated paths regardless of OS:
  use path.Join in List, and convert at the filesystem boundary in
  fullPath (FromSlash) and fileInfo.Path (ToSlash).
- Move directory fsync and rename into build-tagged helpers
  (driver_unix.go / driver_windows.go). Directory fsync is a no-op on
  Windows, and rename falls back to remove-then-rename when the
  destination already exists.
- Harden PutContent error handling: keep a defensive deferred Close,
  make fileWriter.Close and Cancel idempotent, and rely on Cancel to
  clean up the temporary file on error.
